requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2017.7.27.1      # via requests
chardet==3.0.4            # via requests
coreapi==2.3.3
coreschema==0.0.4         # via coreapi
django-filter==1.1.0
django==1.11.6
djangorestframework==3.7.1
idna==2.6                 # via requests
itypes==1.1.0             # via coreapi
jinja2==2.9.6             # via coreschema
markupsafe==1.0           # via jinja2
psycopg2==2.7.3.1
pytz==2017.2              # via django
requests==2.18.4          # via coreapi
uritemplate==3.0.0        # via coreapi
urllib3==1.22             # via requests
